feat: add device controls section for connected devices
Add output sliders (value + position-with-duration) and sensor input
controls to the device detail page, rendered only when the device is
connected. Each output cubit gets its own BlocBuilder for efficient
per-slider rebuilds. Includes battery percentage indicator and sensor
read buttons.
